Renamed IERC20.json as IERC20+IHRC.json, IERC721.json as IERC721+IHRC.json.
