+++ tests/unit/token/TokenDetails.spec.ts
@@ -76,8 +76,8 @@ describe("TokenDetails.vue", () => {
         mock.onGet(matcher3).reply(200, []);
         const matcher4 = "/api/v1/transactions"
         mock.onGet(matcher4).reply(200, SAMPLE_TRANSACTIONS);
-        const abi = require('../../../public/abi/IERC20.json')
-        const matcher6 = "http://localhost:3000/abi/IERC20.json"
+        const abi = require('../../../public/abi/IERC20+IHRC.json')
+        const matcher6 = "http://localhost:3000/abi/IERC20+IHRC.json"
         mock.onGet(matcher6).reply(200, abi)
 
         const wrapper = mount(TokenDetails, {
@@ -146,8 +146,8 @@ describe("TokenDetails.vue", () => {
         mock.onGet(matcher2).reply(200, SAMPLE_NFTS);
         const matcher3 = "/api/v1/contracts/" + testTokenId + "/results"
         mock.onGet(matcher3).reply(200, []);
-        const abi = require('../../../public/abi/IERC721.json')
-        const matcher6 = "http://localhost:3000/abi/IERC721.json"
+        const abi = require('../../../public/abi/IERC721+IHRC.json')
+        const matcher6 = "http://localhost:3000/abi/IERC721+IHRC.json"
         mock.onGet(matcher6).reply(200, abi)
 
         const wrapper = mount(TokenDetails, {
@@ -210,8 +210,8 @@ describe("TokenDetails.vue", () => {
         mock.onGet(matcher2).reply(200, SAMPLE_NFTS);
         let matcher3 = "/api/v1/contracts/" + testTokenId + "/results"
         mock.onGet(matcher3).reply(200, []);
-        const abi = require('../../../public/abi/IERC721.json')
-        const matcher61 = "http://localhost:3000/abi/IERC721.json"
+        const abi = require('../../../public/abi/IERC721+IHRC.json')
+        const matcher61 = "http://localhost:3000/abi/IERC721+IHRC.json"
         mock.onGet(matcher61).reply(200, abi)
 
         const wrapper = mount(TokenDetails, {
@@ -246,8 +246,8 @@ describe("TokenDetails.vue", () => {
         mock.onGet(matcher3).reply(200, []);
         const matcher4 = "/api/v1/tokens/" + testTokenId + "/nfts"
         mock.onGet(matcher4).reply(200, SAMPLE_NFTS);
-        const abi2 = require('../../../public/abi/IERC20.json')
-        const matcher62 = "http://localhost:3000/abi/IERC20.json"
+        const abi2 = require('../../../public/abi/IERC20+IHRC.json')
+        const matcher62 = "http://localhost:3000/abi/IERC20+IHRC.json"
         mock.onGet(matcher62).reply(200, abi2)
 
         await wrapper.setProps({
@@ -309,8 +309,8 @@ describe("TokenDetails.vue", () => {
         mock.onGet(matcher2).reply(200, SAMPLE_NFTS);
         const matcher3 = "/api/v1/contracts/" + testTokenId + "/results"
         mock.onGet(matcher3).reply(200, []);
-        const abi = require('../../../public/abi/IERC721.json')
-        const matcher61 = "http://localhost:3000/abi/IERC721.json"
+        const abi = require('../../../public/abi/IERC721+IHRC.json')
+        const matcher61 = "http://localhost:3000/abi/IERC721+IHRC.json"
         mock.onGet(matcher61).reply(200, abi)
 
         const wrapper = mount(TokenDetails, {
@@ -362,8 +362,8 @@ describe("TokenDetails.vue", () => {
         mock.onGet(matcher2).reply(200, SAMPLE_NFTS);
         const matcher3 = "/api/v1/contracts/" + testTokenId + "/results"
         mock.onGet(matcher3).reply(200, []);
-        const abi = require('../../../public/abi/IERC721.json')
-        const matcher61 = "http://localhost:3000/abi/IERC721.json"
+        const abi = require('../../../public/abi/IERC721+IHRC.json')
+        const matcher61 = "http://localhost:3000/abi/IERC721+IHRC.json"
         mock.onGet(matcher61).reply(200, abi)
 
         const wrapper = mount(TokenDetails, {
@@ -417,8 +417,8 @@ describe("TokenDetails.vue", () => {
         mock.onGet(matcher2).reply(200, SAMPLE_NFTS);
         const matcher3 = "/api/v1/contracts/" + testTokenId + "/results"
         mock.onGet(matcher3).reply(200, []);
-        const abi = require('../../../public/abi/IERC721.json')
-        const matcher61 = "http://localhost:3000/abi/IERC721.json"
+        const abi = require('../../../public/abi/IERC721+IHRC.json')
+        const matcher61 = "http://localhost:3000/abi/IERC721+IHRC.json"
         mock.onGet(matcher61).reply(200, abi)
 
         const wrapper = mount(TokenDetails, {
@@ -459,8 +459,8 @@ describe("TokenDetails.vue", () => {
         mock.onGet(matcher3).reply(200, []);
         const matcher4 = "/api/v1/network/exchangerate"
         mock.onGet(matcher4).reply(200, SAMPLE_NETWORK_EXCHANGERATE);
-        const abi = require('../../../public/abi/IERC20.json')
-        const matcher61 = "http://localhost:3000/abi/IERC20.json"
+        const abi = require('../../../public/abi/IERC20+IHRC.json')
+        const matcher61 = "http://localhost:3000/abi/IERC20+IHRC.json"
         mock.onGet(matcher61).reply(200, abi)
 
         const wrapper = mount(TokenDetails, {
@@ -520,8 +520,8 @@ describe("TokenDetails.vue", () => {
         mock.onGet(matcher3).reply(200, []);
         const matcher4 = "/api/v1/network/exchangerate"
         mock.onGet(matcher4).reply(200, SAMPLE_NETWORK_EXCHANGERATE);
-        const abi = require('../../../public/abi/IERC721.json')
-        const matcher61 = "http://localhost:3000/abi/IERC721.json"
+        const abi = require('../../../public/abi/IERC721+IHRC.json')
+        const matcher61 = "http://localhost:3000/abi/IERC721+IHRC.json"
         mock.onGet(matcher61).reply(200, abi)
 
         const wrapper = mount(TokenDetails, {
@@ -579,8 +579,8 @@ describe("TokenDetails.vue", () => {
         mock.onGet(matcher2).reply(200, SAMPLE_NFTS);
         const matcher3 = "/api/v1/contracts/" + testTokenId + "/results"
         mock.onGet(matcher3).reply(200, []);
-        const abi = require('../../../public/abi/IERC721.json')
-        const matcher61 = "http://localhost:3000/abi/IERC721.json"
+        const abi = require('../../../public/abi/IERC721+IHRC.json')
+        const matcher61 = "http://localhost:3000/abi/IERC721+IHRC.json"
         mock.onGet(matcher61).reply(200, abi)
 
         const wrapper = mount(TokenDetails, {
