+++ tests/unit/transaction/TransactionDetails.spec.ts
@@ -1073,8 +1073,8 @@ describe("TransactionDetails.vue", () => {
             actions: [action], "links": {"next": null}
         })
 
-        const abi = require('../../../public/abi/IERC20.json')
-        const matcher6 = "http://localhost:3000/abi/IERC20.json"
+        const abi = require('../../../public/abi/IERC20+IHRC.json')
+        const matcher6 = "http://localhost:3000/abi/IERC20+IHRC.json"
         mock.onGet(matcher6).reply(200, abi)
 
         const wrapper = mount(TransactionDetails, {
